Move scabbard state TOML conversion
This change moves the conversion of of the TOML version of ScabbardState
to the config value to the `config::toml` module. The
`config::scabbard_state` module should not have any dependencies on
specific formats.
